Fix 64-bit Mac warnings/memory corruption
git-svn-id: http://skia.googlecode.com/svn/trunk@569 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ src/core/SkPaint.cpp
@@ -717,7 +717,7 @@ typedef int64_t Sk48Dot16;
 
 #ifdef SK_SCALAR_IS_FLOAT
     static inline float Sk48Dot16ToScalar(Sk48Dot16 x) {
-        return x * 1.5258789e-5f;   // x * (1 / 65536.0f)
+        return (float) (x * 1.5258789e-5);   // x * (1 / 65536.0f)
     }
 #else
     static inline SkFixed Sk48Dot16ToScalar(Sk48Dot16 x) {
